refactor: useGetValidatedInfo export
